moving zip controller out of API namespace, to use things like flash notices
