+++ config/routes.rb
@@ -18,10 +18,11 @@ Rails.application.routes.draw do
     resources :insta_upload
     
     resources :results
-
+    
+    get :zip_images, to: 'image_zip#zip_images'
     # post '/sites/csv/results', to: 'csv#results', as: :results
   end
-
+  
   # API  
   namespace :api do
     namespace :v1 do
@@ -33,7 +34,6 @@ Rails.application.routes.draw do
       get :site_specific_tags_report, to: 'csv#site_specific_tags'
       get :all_tags_report, to: 'csv#tags_report'
       get :image_quality_report, to: 'csv#image_quality'
-      get :zip_images, to: 'image_zip#zip_images'
       get :download_zip, to: 'image_zip#download_zip'
     end
   end
